requirements.txt
telethon>=1.42.0
aiohttp>=3.8.4
python-dotenv>=1.0.0
